refactor main pipeline into a single fxn w/ a testable return val
